Finishing Part of the works, and big stimulate in sudo codes
During the work, I found that there are many function which otherwise
will be hard to achieve, so I add a lot of new variable and method to
make it OK. I try more and more to achieve administer functions in the
main function. Maybe this will be a new route so solve the problems.